Move getLastUsedDsspResNumOfDsspFile and req class vars to DSSP Parser
